Pylint ingestion models
