+++ ingestion/src/metadata/ingestion/models/profile_data.py
@@ -8,7 +8,9 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
+"""
+Custom models for profile models
+"""
 from pydantic import BaseModel
 
 from metadata.generated.schema.api.data.createTableProfile import (
