+++ ingestion/src/metadata/ingestion/models/tests_data.py
@@ -8,7 +8,9 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
+"""
+Custom pydantic models for tests suites and requests
+"""
 
 from pydantic import BaseModel
