+++ ingestion/src/metadata/ingestion/models/user.py
@@ -8,7 +8,9 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
+"""
+Custom class for User data
+"""
 from typing import List, Optional
 
 from pydantic.main import BaseModel
