+++ ingestion/src/metadata/ingestion/sink/elasticsearch.py
@@ -44,7 +44,7 @@ from metadata.generated.schema.entity.teams.user import User
 from metadata.generated.schema.type.entityReference import EntityReference
 from metadata.ingestion.api.common import Entity
 from metadata.ingestion.api.sink import Sink, SinkStatus
-from metadata.ingestion.models.table_metadata import (
+from metadata.ingestion.models.es_documents import (
     DashboardESDocument,
     ESEntityReference,
     GlossaryTermESDocument,
